+++ core/src/main/kotlin/common/dynamiclink/api/FirebaseDynamicLinkApi.kt
@@ -1,10 +1,8 @@
-package com.wafflestudio.snu4t.dynamiclink.api
+package com.wafflestudio.snu4t.common.dynamiclink.api
 
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import org.springframework.http.HttpHeaders
-import org.springframework.http.MediaType
 import org.springframework.http.client.reactive.ReactorClientHttpConnector
 import org.springframework.web.reactive.function.client.WebClient
 import reactor.netty.http.client.HttpClient
@@ -24,8 +22,6 @@ class FirebaseDynamicLinkApiConfig(
         return WebClient.builder()
             .clientConnector(ReactorClientHttpConnector(httpClient))
             .baseUrl(FIREBASE_DYNAMIC_LINK_BASE_URL)
-            .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-            .defaultHeader(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
             .build().let(::FirebaseDynamicLinkApi)
     }
 }
