+++ core/src/main/kotlin/common/dynamiclink/dto/DynamicLinkRequest.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.snu4t.dynamiclink.data
+package com.wafflestudio.snu4t.common.dynamiclink.data
 
 data class DynamicLinkRequest(
     val dynamicLinkInfo: DynamicLinkInfo,
